Relocated FatJarDetector Utility to generator/java-exec pacakge
