+++ generator/java-exec/src/test/resources/fatjar-simple/test.jar
