+++ generator/openliberty/src/main/java/io/fabric8/generator/openliberty/OpenLibertyGenerator.java
@@ -28,7 +28,7 @@ import org.apache.maven.plugins.assembly.model.DependencySet;
 import org.apache.maven.plugins.assembly.model.FileSet;
 import org.apache.maven.project.MavenProject;
 
-import io.fabric8.maven.core.util.FatJarDetector;
+import io.fabric8.maven.generator.javaexec.FatJarDetector;
 import io.fabric8.maven.core.util.MavenUtil;
 import io.fabric8.maven.docker.config.AssemblyConfiguration;
 import io.fabric8.maven.docker.config.ImageConfiguration;
